Commented relocation of GNPs that are separated less than the van der Waals distance
I commented the part of the code that moves a GNP when it is separated less than the van der Waals distance from another GNP. This, of course, includes touching. This is done to save time during generation since the cases when a GNP needs to be moved will be reduced.