Corrections to GJK for distance calculation
Made the following corrections for the GJK for distance:

1) Support point is added to simplex only when it is not contained in it.
2) Distance from support point to simples when the simples is an edge was corrected.
3) When determining the Voroni region where the origin is located when the simplex is a tetrahedron, a dot product was being used twice in the same condition. One of those dot products was changed.
4) Added and corrected comments for debugging.

Another important change that affects the results of the GJK for distance is the value of the variable "Zero", which is the value that is considered to be numerically zero. It was changed from 1.0E-8 to 1.0E-10. This worked well in the tests I run in Eclipse.