Corrected logic for detemining if a GNP-GNP junction was inside the sample
When the junction points of a GNP-GNP junction are found, it is also determined whether both points are inside the sample or not. In case one of them is not inside the sample, the junction is ignored and the GNPs are not considered to be in contact.

To determine whether the junction points are inside the sample or not there is a flag. When both junction points are inside the sample this flag is set to true. However, this flag was not being activated set to true when hte junction points were found to be inside the sample. This condition was added.