Added missing conditions to GJK for distance
In the GJK for distance, when finding the Voroni region where the origin lies in the case the simplex has size 3, some conditions were missing. When deciding if the origin lies in the Voroni region of an edge I added two conditions that ensured the origin is in the Voroni region of the edge. Without this conditions, the orig lies in the hlafplane perpendicular to the simplex that contains a given edge.

I also further reduced the Zero variable to 1e-12 since the previous value of 1e-10 was causing the GJK for distance to reach the maximum number of iterations.