Added error message to EPA
Due to possible numerical errors, the GJK for distance may terminate with the interpenetration condition when it should not. This will cause the EPA to try to access an element of an empty vector. Of course, such vector is not empty when GNPs are actually interpenetrating each other. Because of this, now I added an error message when that vector is empty.