Corrected traingulation
I had changed the supertriangle to a supertetrahedron to generate the triangulation in 3D for GNP junction points. However, the whole algorithm needs to change to actually be fully 3D. That is, tetrahedrons need to be constracted instead of triangles. So, I reverted back to using a supertriangle. In addition I added an error message to detect when a triangulation using the Bowyer-Watson algorithm is not being constructed.

I found that one of the vertices of the supertetrahedron was not initialized correctly. So I also corrected this.

I also found that the edges of the final triangulation were being ignored. This because I was ignoring all edges that were repeated. However, what I needed to do was ignore the repetitions. Now the triangulation is done properly.

To help debugging I added some comments and added the possibility to sport vtk files for triangles and the supertetrahedron.