Moved location of HK76 execution for GNP networks
When only GNPs are generated, it may happen that the junction between GNPs is outside the sample. In such case, junctions are ignored. However, GNPs were being connected via the Hoshen-Kopelman algorithm (HK76). Thus GNPs that were connected to no other GNP were part of a percolated cluster. This caused having zeros in the diagonal of the stiffness matrix, which lead to division by zero when applying the pre-conditioner.

Now, the HK76 is performed only if the junction points are both inside the sample.

Also, the diagonal of the stiffness matrix is scanned to find zeros. If any zeros are found, an error message is generated, all nodes with zero are displayed in the output file and the function terminates with an error.