Corrected initial guess for CNT nodes
The calculation of the initial guess only worked for CNT nodes. So, when only GNPs are present there is an error because the code tries to access an element of the vector of CNTs in the cluster, which is empty since there are no CNTs. Thus, I added a conditio to check that there are CNT clusters and the cluster is not empty in order to calculate the initial guess.

I also encapsulated some parts of the code into functions.