Clear triangulations from second iteration onwards
Triangulations are cleared for all iterations except the first. This caused having more triangulation edges than needed with nodes that were not there anymore in smaller observation windows. Of course, this led to trying to access elements out of the reach of a vector.

I also encapsulated some code into a function and added some comments.