Added function to clear triangulations
Added a function that, in the case of multiple percolated clusters, clears the traingulations of the previous cluster. Triangulations of the previous cluster could intefere with the calculations on the current cluster if a GNP happens to be in the boundary of the window.

Also, I modified the way I was clearing the triangulations at the biginning of the electrical analysis. Triangulations of all GNPs in a cluster used to be the ones that were cleared. Now, triangulations of all GNPs are cleared.

Also added some comments for debugging.