Comparison with zero in GJK for distance
In the GJK for distance, whenever a comparison is made against zero to determine the sign of a number, now I use 0.0. In the cases when the comparison against zero is to determine two numbers are the same (or practically the same) I use the variable Zero. At this point the variable Zero is 1e-12.