Reconstruction of GNP completely inside the sample
Implemented the functions that reconstruct a GNP that is completely inside the sample.

For this I also had to create a new function in class Generate_Network that generates a rotation matrix from a given direction. In this way I can use the code that I already wrote instead of copying it and having duplicated code.