Corrected comparisons with zero and updated default constructors for geometry objects
Corrected some comparisons with zero where a "0.0" was used instead of the variable "Zero".

I updated the default constructors of Point_3D and Plane_3D so that variables are initialized when using these constructors.

I also inverted the usage of the virtual_plane flag. It used to be false for a virtual plane and ture for a real plane. Given the name of the flag, I would expect it to be true for a virtual plane. So I changed its definition. Now virtual_plane is true for a virtual plane and flase for a real (non-virtual plane).