Corrected adjustment of overlapping regions for contact grid
When calculating the number of overlapping subregions for the contact grid I added a 1 in case numerical issues resulted in going avobe the calculated number of regions. However, if the size of the subregions needed to be adjusted, this 1 was ignored. The resulted in calculating subregions outside the maximum number of subregions. I also noticed that I was calculating the adjusted size of subregions in a non-efficiente manner. So I corrected this two issues.

I also modified some comments to add more information when an error was found in the function that assigns a CNT point to the overlapping subregions.