Corrected calculation of point to line
The way the calculation of a point to a line was done was actually incorrect. This was only implemented in the Line_3D class. So I corrected the implementation and also added an implementation to the Point_3D class. In this way I do not need to create a line object and can directly use a function from the Point_3D class.

I also corrected the operators == and != to use the Zero variable.