Incorrect variable used
There was a variable being used that hay nothing to do with the code. This was added when editing the code in Xcode, so probably it was an Xcode variable which explains why I did not get any errors when using it. It turns out the variable was not even declared. So here I corrected that.