Corrections to GNP reconstruction
Added an error message in case there are less than three GNP vertices inside the sample when reading the GNP data from csv.

Added a missing space in the vertices line for exporting visualization files for points.

Changed the logic conditions to identify case 4 for reconstructing a GNP.

Corrected update of two vertices in case 2 for reconstructing a GNP.