Updated the way visualization files are generated when reading network from file
When reading a network from file, the only way visualization files were generated was with all CNTs or GNPs in a single file. Other options, though available in the input file, were not available in the implementation of the functions when reading a network from a file.

I changed this and now a single CNT or GNP can be generated on a single visualization file. However, the option to output CNTs as genenarated before trimming will be ignored as the information necessary to generate such file is not available from the csv or binary files.