HK76 for penetrating GNPs
Now the Hoshen-Kopelman algorithm works for GNPs that are penetrating each other. To do this, one GNP is moved and the junction points are found with this GNP. Also, to move the GNP a temporary variable is used in order to not modifying the original network.

I also noticed that a couple of variables were not being used in the HK76 for GNPs only. These variables sorted the GNP numbers with GNPa with the lowest one and GNPb with the largest one.

Also added the GNP numbers to some error messages.