Correction in HK76 for penetrating GNPs
When using the Hoshen-Kopelman algorithm for only GNPs, I made a correction to the contact point of the GNP that is moved when penetrating another GNP. The calculated point is on the moved GNP, which is saved into a temporary variable. So the correction consisted in moving the point back into the original position (in which it penetrates another GNP).

I also noticed that the GNP point numbers were not assigned properly. There is point Pa for GNP A and point Pb for GNP B. Pa is added first to the vector of GNP points and then Pb. Thus, Pb is equal to the number of elements in the GNP point vector minus one. Similarly, Pa is equal to the number of elements in the GNP point vector minus two. However, these point number were calculated the other way around. Now the point numbers are obteained properly.