Changed the way displacements are applied to GNPs
I realized thar when applying the displacements to GNPs, after frame 1 the method I was using is not valid. This beacuse displacements are applied to the reconstructed GNP, not the original GNP nor the GNP with the displacements in the previus frame.

Thus, in the case of GNPs a copy of the original GNPs are geenrated. And the displacements of a given frame are applied to the original GNP. In this way the previous frame is not needed for GNPs. It is needed, of course, for CNTs and the sample.