Changed value of Zero variable
Found an example that resulted in the GJK finding penetration of GNPs when in reality there was no penetration. This was due to numerical error in the termination criteria when using a Zero variable of 1e-12. The Zero variable is changed to 1e-10. In the future, if needed, the Zero variable will be modified locally.

I also added an error message for the case when the simplex that is used as input to the EPA algorithm is not a tetrahedron. The EPA needs a tetrahedron to run since it is the only simplex that encloses a point in 3D.

Added and modified some comments for debugging.