Observation window extraction for GNPs with displacements
When reading displacements from an Abaqus database, GNPs are reconstructed by finding the parallelepiped that encloses the vertices inside the sample. This causes the GNP points at the boundary to change coordinates. These GNP boundary points are not updated since they are only calculated when extracting the observation window. This is contrary to CNTs, where CNT points at the boundary are actually part of the CNT and are updated when applying displacements to the CNT points. Thus, after applying displacements to the GNP vertices the procedure to extract an observation window should be performed but only to GNPs. This is done in this update. Before doing so, the vectors generated during such procedure are emptied.

In addition, the vector points_gnp needs to be cleared before the Hoshen-Kopelman algorith. Otherwise there will repeated GNP points. So this is also done in this update.

I changed some variable usage that seemed a bit overcomplicated in the Hoshen-Kopelman algorithm.