Correction to GNP reconstruction case 1
I realized that the GNP reconstruction case 1 was not considering all possibilities. I was only considering completing an adjacent edge to the only edge inside the sample and then go to case 2 (two consecutive edges inside the sample). However, it could happen than there are only three vertices inside the sample and the third vertex is across the reference vertex. I was not considering this case. So in this update I also search for this case and now from case 1 reconstruction could go to case 2 or case 4.