Corrections when generating visualization files for isolated particles
When exporting visualization files for isolated particles, the functions that generated the vtk files were called for both CNTs and GNPs without checking first that there were indeed CNTs or GNPs. This caused a runtime error as an element was being accessed in an empty vector. Now I check if there are CNTs or GNPs to export visualization files for isolated CNTs or GNPs, respectively.