Moved declaration of vector for GNP points and structure to reduce memory usage
For the case of the resistor network application, I realized that the vectors points_gnp and structure_gnp were declared before looping on the number of observation windows.

In the case of the structure vector, it was reset at every iteration when extracting an observation window. However, the vector of GNP points increased at every iteration, since these vector at iteration i contained all poits generated from iteration 0 to iteration i. Thus, by moving the declaration of these vectors to the beginning of the loop over the observation windows, all points generated on a single iteration are discarded for the next iteration. This reduces memory usage.