Simplified a calculation of distance from point to segment
Changed the way a function in the class Generate_Network calculates the distance from a point to a segment defined by two points. Now I use the function that I defined in the Point_3D class. Also, the point in the segment closest is found. This procedure is also simplified.