Corrected GNP junction points on the same location
I found that in some cases junction points were repeated. That is, a GNP had a junction point that was shared with two other GNPs. However, this shared point was added twice into the vector of points. Which resulted in a GNP resistor with zero length and, thus, it had infinite resistance.

This happened with GNPs taht were penetrating each other but the penetrations were outside the sample. In principle these penetrations were ignored since they were outside the sample and did not affect the calculations nor the model in Abaqus. However, due to the repetition of junction points penetrations need to be avoided also outside the sample.

Thus, I modified the algorithm that added GNPs to the overlapping subregions to detect GNP penetrations. Now there are subregions outside the sample to add GNPs that are partially outside the sample and detect their interpnetrations with other GNPs. Also, I simplified the way GNPs were added into a subregion and made it more similar to the case of CNT points.

I changed how the discretization of GNPs was done. I was using the size of the subregions to make the discretization. However, I realized that for the case of only GNPs it would be more efficiente to have large subregions (around half the GNP side length) and have several points along the GNP side length (around 20). I found this largely reduced the computational time. Proabbly changing the way the GNPs are added to subregions also contributed in reducing computational time.

When calculating the junction resistance I added a condition to check the length of the resistor. This condition terminates the execution (returns 0) with an error when the length is less than the variable Zero.

I also added some lines of code to set the overlapping of CNTs (or GNPs) to zero when only GNPs (or CNTs, respectively) are generated. This, so that one overlapping does not interfere with the other one.

I modified and added some comments for debugging.