Corrected comparison with zero in EPA
In the EPA, some comparisons with zero are needed. In the code I compare with the variable Zero. this caused some errors that resulted in not being able to construct the polytope needed to calculate the distance that penetrating GNPs needed to be moved to avoid penetrations. From the errors found in the GJK algorithm, it proved more useful to use 0.0 instead of the variable Zero wihen comparing with zero. Thus, to avoid the errors in EPA I changed some comparisons with the variable Zero to comparisons with 0.0.