Missing update when adding GNPs to subregion
I made a correction when adding GNPs to a subregion that was needed to avoid interpenetration outside the sample, which in turn caused junction points to be repeated, which in turn caused resistors with length zero and infinite resistance.

The same correction was needed in the other part of the code where GNPs where added into subregions. That is, in the Contact_grid class. This change, sets the minimum number of descretization points of a GNP to 20 along the x- and y-directions. This is specially needed when having large subregions compared to what is usually needed for CNTs.