Modified the way CNT points are added into subregions during generation
I modified the way CNT points are added into subregions (used to search for interpenetrations) so that it matches the changes that I did for GNPs in contact grid. This will reduce in some cases the number of subregions a CNT point belongs to. As a consequence I expect the code to run a bit faster since penetrations not be searched for in less subregions (in some cases).

Also added some comments for debugging.