Vector of all resistors for debugging
Added a vector as a global variable in the class for the DEA. This vector is used to save all resistors and print a file with all the resistor values sorted for debugging. All new lines are comments as this change is inteded for debugging only, they're useless at runtime.