Reduction in number of operations to calculate junction resistance
Modified the way the junction resistance is calculated. I noticed that many of the operations performed to calculate the junction resistance cosist of quantities that are constant during the whole simulation. Thus, some values can be pre-calculated and be used when needed instead of calculating the same quantity several times. Thus, in this update I identified two constants that simplify and make more efficient the calculation of junction resistance.