Corrected mv command for visualization files
When exporting visualization files for percolated clusters there was a message on the cmd that mv found no files with name isolated_*.vtk. After checking the code I noticed that I tried to move twice the files with name isolated_*.vtk and thus the error appeared. For this update I just removed the repetition of isolated_*.vtk.