Corrected the way junction resistance is calculated
The junction distance is saved in a junction object. However, I was calculating it wavery time. Not only is this not efficient, but it also lead to extremely high junction resistances. This because when reading positions from Aabqus, distances between junction points may be too large due to interpenetrations when reconstructing GNPs. This is the case specially when there are large deformations.

Thus, now the junction resistance is calculated using the distance stored in the junction object.

Also modified and added some comments for debugging.