Discard repeated points in GNP junctions
Since there are interpenetrations when using only GNPs and reading the network from Abaqus due to the way GNPs are reconstructed, sometimes a GNP has two junction points that are actually the same.

So I added a function to check for repeated points when there are interpenetrations. For this I defined a tolerance that determines when to consider two points to be the same when adding a junction. This tolerance is not the variable Zero so that points may e a bit far but condensed into one. For now this tolerace is 1.5 times th thickness of the GNP. In addition I added some functions that add some clarity to the code.

I also modified some comments for debugging.