Flag for exporting visualization files for generated GNPs not working properly
The flag for exporting generated nanoparticles with value 3 was supposed to only have different effect for CNTs but still allow exporting visualization files for GNPs. However, files for GNPs were not being exported. This is because the value 3 for the flag was not considered when exporting generated GNPs. Thus, in this update the value of 3 also generates visualization files for generated GNPs.