Correction on tunneling distance
The old function that I had to calculate the junction resistance, adjusted the distance if a CNT was part of the junction. After I modified the function I deleted that adjusment. However, that adjusment is still needed. The adjusment was done in the Hoshen-Kopelman algorithm when saving a junction. Here, the actual tunneling distance is calculated instead of the distance from a CNT point to either another CNT point or a GNP.

A second adjustment to the tunneling distance was to set a minimum values. This beacuase Abaques actually allows interpenetration of instances with embedded elements. Thus, distances between points are less than the sum of the CNT radii which leads to a negative tunneling distance. Thus, now the minimum tunneling distance is enforced as the van der Waals distance.

Added a couple of lines to adjust the displacements applied to the RVE in case some CNT elements from Abaqus lie outside of it. This is needed when I increase the size of the RVE in Abaqus to have all CNT elements lie inside the RVE and run the simulation. The lines I added are commented so they don't affect the normal flow of the code but incase they are needed they'll have to be uncommented.

Corrected some typos in comments and deleted some comments that made reference to old code that is not there anymore. Added more comments for debugging.