Overlap for CNTs not set for mixed particles
The overlap for CNTs was only being calculated for the CNT deposit and only CNTs. However, it was not calculated for mixed particles althoug there were CNTs present. So, now it is calculated for mixed particles and updated the comment to reflect that this needs to be calculated whenever there are CNTs.