Corrected interpenetration of CNTs over GNPs
I implemented some code to deal with mixed interpenetrations while considering overlapping subregions were defined in a certain way. However, it turns out I got confused and the overlapping subregions were defined differently. This caused the code to ignore some interpenetrations. I revised how subregions are defined and made the necessary changes.

I found another bug related to interpenetrations within CNTs. In this bug, there could be the case when a point in the CNT currently being generated is outside the sample and interpenetrating a point from another CNT that is inside the sample. The point generated next could be inside the sample and also penetrate the point from another CNT. In such case, I cannot calculate the new point's position analitically because the triangle inequality will not be satisfied if the point before the new point is also penetrating the other point. This causes the code to attempt to take the square root of a negative number. to solve this, I defined a non-penetration layer (NP layer). This layer extends a bit more than twice the maximum CNT radius from the sample. Thus, in the sample and NP layer there will be no interpenetrations. This will eliminate interpenetrations at the boundary which used to happen as in the example mentioned earlier. In addition, I also check for the case when in the newly generated CNT the previous point is outside and the new point is inside. In such a case the point is moved wihtout presetving CNT segment length as this segment will be outside the sample.

Added and corrected some comments.