Points on the sampe plane before doing triangulation
I found that some times a GNP point was not part of the triangulation for a given GNP. This seemed to be due to points not being on the same plane, since the triangulation is done using a 2D algorithm. Thus, I moved all points to the midplane of the GNP, which is also where the supertriangle lies. So far this has solved this issue.

So for this change, I create a new vector with the points on the plane and also a structure vector with numbers 0 to n-1, where n is the number of points to triangulate on a GNP. Then the triangulation is done. Edges will contain vertices with numbers from 0 to n-1. Finally, numbers 0 to n-1 are changed to the original numbers.

Also corrected and added comments for debugging.