Scalaing of resistances (internal only)
Added some scaling when calculating resistances in order to reduce numerical errors when solving the system of equations for the electrical circuit. All resistances are now in MOhms. When the equivalent resistance is calculated, the inverse scaling is applied so that the user does not see the scaling going on.

I also modified some indentations and commented some lines of code that I forgot to comment.