Removed scaling, increased maximum iterations for CG
There seemed to be some numerical issues were there were negative resistances and the CG did not converged. I thought these were numerical errors due to poor scaling or not using an appropriate preconditioner. Thus I first implemented scaling of resistances, which in some cases it seemed to work. However, in other cases it was not enough.

At the end, I realized that the CG just needed more iterations to converge. So I increased the maximum nmber of iterations for the CG before terminating. Also, I commented the lines of code for scaling in case they are used in the future.