Changed the way the backbone for GNPs is determined
The GNP backbone was obtained by calculating the current passing through each GNP resistor and if all reisstors had current lower than the zero current then the GNP was dead.

However, I realized that this is not how it should be done. The current through a junction is split when going through the GNP resistors. Thus, I should check if the current though the junctions is above the zero current. If there are junctions and GNP resistors left in the GNP, then the GNP is part of the backbone. So I implemented this change and had to had more data structures. I also had to modify other classes besides the one that finds the backbone.

An important change is that, after determining percolation, all boundary points that are not in relevant boundaries are deleted. A point is in a relevant boundary if it is in a boundary connected by the backbone.