Update for the case when a CNT segment is larger than GNP thickness
In this update I consider the case when a CNT segment may be larger than the GNP thickness. When this happens, the newly generated point my end up in a position where the new CNT segment traverses the GNP. Although this type of penetration may be handled by the code, It may take many iterations or result in rejecting the CNT. Thus, I check for this case and rotate the CNT lenght in such a way that may helo the code in acoid the interpenetration in the next iteration.