New function for moving CNT point or rotating CNT segment (mixed particles)
I created a new function when checking for mixed interpenetration. This new function moves the newly generated CNT point or the last CNT segment depending on whether the CNT point is a seed or not.

I added this function since I nedeed to make the choice (moving or rotating) in two parts of the function that checks mixed interpenetration. So instead of copying the code, I created the function and called it twice.

I also modified some comments for debugging.