Corrected comments and text sent to the output file
In the Input_Reader class made a couple of changes. Commented a couple of hout commands that were not necessary to be in the output file. Also, it was missing an update in the criterion for the gnp_geo object.

I also separated in two lines a message regarding the amount of GNPs generated.