Modified function that moves a CNT point generated inside a GNP
During generation of mixed particles, a CNT point may be generated inside a GNP. I had made a function to deal with this case and move the newly generated CNT point to avoid interpenetrations. However, the way that function worked, did not preserved the length of the CNT segment. This caused an error in the code when rotating a CNT segment.

Thus, I changed this function and now the CNT segment is preserved.

Also changed some comments.