Had not implemented the non-penetration layer for CNTs only properly
For CNTs only I actually needed to use the function that calculates the subregion number with layer since I am also using here the non-penetration layer. Thus, I changed the name of the function so that it could be more representative of only CNTs and mixed particles and now use it in both cases.

I also changed the name of a function that had a typo and deleted some comments that are not representative of the code in its actual form.