Incorrect assignment of CNT points to overlapping subregions
Realized that for the generation of CNT points and to avoid interpenetrations, they need to be assigned to all the overlapping subregions. Only when all points are generated and to search for contacts a simplification can be made where a point is not assigned to al the overlapping subregions.

Thus in this change, for the generation of GNPs, all the generated CNT points are assigned to all overlapping subregions.