Fixed numerical errors when rotating a CNT segment
When a penetrating CNT point was found and it is not a seed, then the CNT segment needs to be rotated. However, after rotation the code still detected interpenetration. This was a numerical error that made the code judge two very similar numbers as similar. So I added some Zeros in order to eliminate this numerical error.