Changed loop for moving penetrating points for only CNTs
Modified the loop that I use to find penetrating points and move them to avoid such penetrations. I changed it from a for-loop to a while-loop to match the loop that I used for mixed particles. In addition, some functions that were called twice in the function, now they are only called once. This reduces repetition of the code.