Completing CNT elements for mixed particles
After adding junctions between CNTs, CNT elements are generated to facilitate the formation of the stiffness matrix. This was only done when only CNTs were generated. However, it is also needed when mixed particles are generated.

So now, when making CNT clusters, CNT elements are completed only when the particle type is not mixed or hybrid. And now, CNT elements are completed after making mixed clusters.