Removing CNT points from element when removing mixed junction
When finding the backbone, junctions that carry a current below a cutoff I0 are removed. However, in the case of mixed particles it could happen that a "mixed" junction carries a current above I0 but the CNT is dead so the CNT point of the "mixed" junction is not present. In the case the junction carries a current below I0, the CNt point is not removed.

So here I added conditions to remove a mixed junction if the CNT is dead or the elements vector does not contain the CNT point. Also, now if the junction is removed, the CNT point is removed from the elements vector.

I also changed some comments for debugging.