Removing CNTs with less points than the minimum number of points in the input file
I found that there could be some CNTs with less points than the minumum number of points indicated in the input file. This happens when the start and/or end indices for a CNT segment move and the number of points is less than the intial number of points in the segment. The initial number of points is the one that is calculated in the code and is also used to determine if a CNT segment is kept or not.

Thus, now, after the CNT segment is added, there is a new check that identifies the CNTs with less points than the minumum indicated. When those CNTs are identified, they are removed from the vector of points.