Flags for exporting CNT and GNP data not working properly
There is a flag for exporting CNT and GNP data into csv or dat files. When this flag is set to 3, both csv and dat files should be exported. However this was not the case because the if-statements were not implemented properly. I corrected this and now flag with value 3 works as it should.