Check odb file exists
I updated the input reader class to check that the odb file or the path to it is valid. That is, I check if the file exists. If the file does not exist as with the name or path indicated in the input file, then the program terminates with an error.