Changed the order in which information is given for GNPs
The number of GNPs inside the sample was given in the output file before the number of generated GNPs. This is in contrast with how it is done for CNTs, where first its given how many CNTs and points were generated, rejected and ignored and then the ones inside the sample.

Now, the number of GNPs inside the sample is given last to match the format in which CNT information is given.