Removed statistics from generated CNTs in the case of mixed fillers
In the case of CNT networks, I count the number of times that there is an overlap of CNT points. I don't do that for the case of mixed fillers. However, the variables to count overlaps are initialized but never used. This causes the message sent with the statistics of the overlaps to indicate 0 overlaps. Since I would need to make more changes to the code to include this statistics, I decieded to remove the variables and do not include those statistics for the case of mixed fillers.