Removed CNTs with no elements
I found that in the case of mixed particles, when removing junctions that are not in the backbone, it may happen that a CNT ends up with only one point in the elements set. In this case, the CNT has no elements.

So now, after all mixed junctions are removed, I check if more CNTs need to be removed from the cluster of CNTs. In addition, when constructing the LM matrix now I check that there is at least one element, i.e., there are at least two elements in the elements set.

I also added some comments for debugging.