Added error message when a CNT point is inside a GNP when looking for mixed junctions
In a simulation when reading positions from an Abaqus database, I found that in some cases some CNT points where inside GNPs. This caused some runtime errors since this was not handled properly. These errors came from some points with the same coordinates when calculating the triangulation. The points had the same coordinates since the function that calculates the point on the GNP returns a point with the coordinates of the origin and flag -1 when a CNT point is inside the GNP. This was never checked. So now I added and if-statement to detect this scenario. This was happening due to the large deformation of the Abaqus simulation. Thus I do not expect this happens in simulations where deformation is smaller.

In addition, I encapsulated some code into a function to reduce the amount of repeated code.