Incresed the number of discretization points for GNPs
When generating GNPs, they are discretized into points so that they can be added to subregions. If this discretization is too coarse, they will not be added to all subregions they occupy. This will result in possible interpenetrations not being detected. Thus, I increased the default number of discrtization points along the local x- and y-directions.

I also added some comments for debugging.