Added flag to ignore an error when a CNT point is inside a GNP after deformation
When deformation is applied, GNPs are reconstructed performing certain approximations. This results in they having slightly larger volume. Adding to this the fact that interpenetrations are possible in the Abaqus simulation due to the use of embedded elements, then it is possible that CNT points end up inside GNPs after deformation is applied. Because of this I decided to ignore when this happens instead of throwing an error.

To ignore the case when the above happens, I created a flag that is initialized to 0 in the constructor of the Hoshen_Kopelman class. Only in the application class for reading from an Abaqus database the falg is set to 1 so that the error is ignored.