Added the possibility to specify relative volume fraction
In the case of mixed particles, now the relative volume fraction or relative weight fraction may be specified for the content of particles. The relative volume fraction is defined as vol_CNT/(vol_CNT + vol_GNP). Here vol_CNT is the volume fraction of CNTs and vol_GNP is the volume fraction of GNPs. Similarly, the relative weight fraction is defined as wt_CNT/(wt_CNT + wt_GNP). Here, wt_CNT is the weight fraction of CNTs and wt_GNP is the weight fraction of GNPs.