Ignoring dead GNP-GNP junctions
I found that some GNP-GNP junctions that were not part of the backbone werenot being removed. I found a case that I needed to consider and added it to the Backbone_Network class. However, there were still some junctions that were not removed and caused errors when filling the stifness matrix. So now, before adding GNP-GNP junction resistors to the stiffness matrix, I first check if both junction points have maps to a node. If both maps exist, then the junction resistor is added. Otherwise the junction is ignored. This is more effective than looking for all the cases of junctions resistors that need to be removed in the Backbone_Network class.

I also added some comments for debugging.