Ignoring csv/dat files when particles are not generated
The way the code was working, if only GNPs were generated but csv/dat files for CNTs were requested in the input file, then csv/dat files were generated. I corrected this and now a warning message is sent to the output file stating that no particles were generated and so no file was generated either.